bugfix(view page):fixed loading for guest users

+++ public/assets/JS/comments.js
@@ -1,7 +1,7 @@
 import  app  from "../../firebase.js";
 import { isLoggedIn, userData } from "./authentication.js";
 import { getFirestore, doc, getDoc,addDoc,collection,deleteDoc,getDocs } from "https://www.gstatic.com/firebasejs/12.2.1/firebase-firestore.js";
-import { postData } from "./view.js";
+import { getPostData, postData } from "./view.js";
 
 const db = getFirestore(app);
 
@@ -110,6 +110,7 @@ export function initializeCommentsBox(){
     };
 };
 export async function renderComments() {
+    await getPostData();
     const comments = await getDocs(collection(db, "posts", postData.uid, "comments"));
     const commentsData = comments.docs.map(doc => ({
         id: doc.id,
@@ -144,27 +145,30 @@ export async function renderComments() {
             `;
         commentsBox.appendChild(commentBox);
 
-        if (comment.userId === userData.uid || userData.role === "owner") {
-            const specialButtons = commentBox.querySelector(".special-buttons");
 
-            const deleteButton = document.createElement("span");
-            deleteButton.classList.add("lios-action-button", "delete-button");
+        if (userData) {
+            if (comment.userId === userData.uid || userData.role === "owner") {
+                const specialButtons = commentBox.querySelector(".special-buttons");
+
+                const deleteButton = document.createElement("span");
+                deleteButton.classList.add("lios-action-button", "delete-button");
 
 
-            deleteButton.innerHTML = //html 
-                `
+                deleteButton.innerHTML = //html 
+                    `
                 <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-trash2-icon lucide-trash-2"><path d="M10 11v6"/><path d="M14 11v6"/><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6"/><path d="M3 6h18"/><path d="M8 6V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"/></svg>
                 `;
-            deleteButton.addEventListener("click", async () => {
-                try {
-                    await deleteDoc(doc(db, "posts", postData.uid, "comments", comment.id));
-                    commentBox.style.display = "none";
-                } catch (error) {
-                    console.error("Failed to delete comment:", error);
-                }
-            });
+                deleteButton.addEventListener("click", async () => {
+                    try {
+                        await deleteDoc(doc(db, "posts", postData.uid, "comments", comment.id));
+                        commentBox.style.display = "none";
+                    } catch (error) {
+                        console.error("Failed to delete comment:", error);
+                    }
+                });
             
-            specialButtons.appendChild(deleteButton);
+                specialButtons.appendChild(deleteButton);
+            };
         };
     });
 }
